refactor the mlp and flow models
