+++ src/mother/ml/models/m_head_utils.py
@@ -1,11 +1,25 @@
 """
 Utility functions for head layers (MLP, Flow, etc.).
 
-This module provides utilities specific to prediction head layers,
-particularly for flow-based probabilistic predictions.
+This module provides utilities shared by the prediction head layers
+(:mod:`mother.ml.models.m_mlp` and :mod:`mother.ml.models.m_flow`): dimension
+auto-detection, dataframe formatting, adaptive MLP sizing, and flow mode /
+uncertainty computation.
 """
 
+from typing import Any, Dict, List
+
+import numpy as np
+import numpy.typing as npt
+import pandas as pd
 import torch
+from optuna import Trial
+from skorch import NeuralNetClassifier
+from skorch.callbacks import Callback
+
+# Default quantiles for the standardised predict_uncertainty interface
+# (mirrors the convention used by CatBoost / TabPFN / RandomForest / NODE).
+DEFAULT_QUANTILES: List[float] = [0.25, 0.5, 0.75]
 
 
 def compute_flow_mode_and_uncertainty(dist, num_samples: int = 100):
@@ -75,3 +89,151 @@ def compute_flow_mode_and_uncertainty(dist, num_samples: int = 100):
         uncertainties = -best_log_probs  # [batch_size]
 
     return mode_predictions, uncertainties
+
+
+def _prepare_for_dataframe(values: npt.NDArray[np.float32]) -> Any:
+    """Convert a prediction / uncertainty array into a column suitable for a DataFrame.
+
+    Single-target arrays are flattened to 1D; multi-target arrays are converted to a
+    list of per-row vectors so each cell holds the full target vector. This mirrors the
+    helper used by the NODE estimators so head outputs share the same layout.
+    """
+    if values is None:
+        return None
+    arr = np.asarray(values)
+    if arr.ndim == 1:
+        return arr
+    if arr.ndim == 2 and arr.shape[1] == 1:
+        return arr.flatten()
+    return [row for row in arr]
+
+
+class DimensionSetter(Callback):
+    """Automatically detect and set ``input_dim`` / ``output_dim`` from training data.
+
+    Runs at ``on_train_begin`` and updates the module's dimension parameters
+    based on the actual data shapes.  This allows users to create standalone
+    MLP / Flow heads without specifying dimensions upfront.
+
+    Detection rules:
+    - **input_dim**: ``X.shape[1]``
+    - **output_dim for classification**: number of unique values in *y*,
+      or ``len(net.classes_)`` if Skorch has already detected them.
+    - **output_dim for regression**: ``y.shape[1]`` if 2-D, else ``1``.
+    """
+
+    def on_train_begin(self, net: Any, X: Any = None, y: Any = None, **kwargs: Any) -> None:
+        """Detect dimensions from data and update module parameters."""
+        # Skip if dimensions are already properly set (not default placeholders).
+        # Use NeuralNet.get_params (via super()) to access module__* keys that
+        # our overridden get_params() strips out.
+        raw_params = super(type(net), net).get_params()
+        current_input_dim = raw_params.get("module__input_dim", 1)
+        current_output_dim = raw_params.get("module__output_dim", 1)
+
+        # Auto-detect only the dimension(s) still at the placeholder value (1). If the
+        # user set BOTH input_dim and output_dim explicitly there is nothing to do;
+        # otherwise we still detect the missing one (e.g. a user who sets input_dim but
+        # leaves output_dim at 1 for multi-class classification).
+        if current_input_dim != 1 and current_output_dim != 1:
+            return  # Both dimensions already set by user
+
+        # Get actual dimensions from data
+        if isinstance(X, pd.DataFrame):
+            input_dim = X.shape[1]
+        elif hasattr(X, "shape"):
+            input_dim = X.shape[1] if len(X.shape) > 1 else 1
+        else:
+            input_dim = len(X[0]) if len(X) > 0 else 1
+
+        # Detect output dimension
+        # For classification, check if we can infer number of classes
+        if y is not None:
+            if hasattr(net, "classes_"):  # Classification task (skorch detected classes)
+                # Number of classes detected by skorch
+                output_dim = len(net.classes_)
+            elif isinstance(net, NeuralNetClassifier):
+                # Classification without pre-detected classes: infer the number of
+                # classes from the unique target values. This "few unique values"
+                # heuristic must NOT be applied to regressors, where an integer-valued
+                # target (e.g. counts) with few unique values would otherwise corrupt
+                # the regressor's output shape.
+                if isinstance(y, pd.DataFrame):
+                    # Multi-label targets are typically shape [n_samples, n_labels]
+                    # with values in {0, 1}. In that case output_dim must be n_labels,
+                    # not the number of unique scalar values.
+                    output_dim = int(y.shape[1]) if y.shape[1] > 1 else int(pd.Series(y.values.ravel()).nunique())
+                elif isinstance(y, pd.Series):
+                    output_dim = int(y.nunique())
+                else:
+                    y_arr = np.asarray(y)
+                    output_dim = (
+                        int(y_arr.shape[1]) if y_arr.ndim > 1 and y_arr.shape[1] > 1 else int(len(np.unique(y_arr)))
+                    )
+            elif hasattr(y, "shape"):
+                # Regression: output dimension is purely shape-based.
+                output_dim = y.shape[1] if len(y.shape) > 1 else 1
+            else:
+                output_dim = 1
+        else:
+            output_dim = 1
+
+        # Update only the placeholder dimensions and force re-initialization so a
+        # user-provided input_dim / output_dim is never overwritten by detection.
+        new_params: Dict[str, int] = {}
+        if current_input_dim == 1:
+            new_params["module__input_dim"] = input_dim
+        if current_output_dim == 1:
+            new_params["module__output_dim"] = output_dim
+        if new_params:
+            net.set_params(**new_params)
+            if net.initialized_:
+                net.initialize()
+
+
+def _suggest_adaptive_hidden_dims(
+    trial: Trial,
+    input_dim: int,
+    *,
+    layers_key: str,
+    width_key: str,
+    max_layers: int = 4,
+) -> List[int]:
+    """Suggest an input-adaptive funnel MLP architecture.
+
+    Shared by the standalone MLP head (``BaseMLPHeadEstimator``) and the flow head's
+    MLP trunk (``BaseFlowHeadEstimator``) so both size their hidden layers identically:
+
+    - **depth** ``[1, max_layers]`` (``layers_key``);
+    - **first-layer width** scaled to the input dimension (``width_key``): from
+      ``max(64, input_dim // 2)`` up to ``input_dim * 2``, step-aligned so Optuna
+      only proposes reachable values;
+    - **funnel**: each subsequent layer halves the previous width, floored at 32.
+
+    Args:
+        trial: Optuna trial used to suggest the depth and first-layer width.
+        input_dim: Number of input features (drives the width scaling).
+        layers_key: Full trial parameter name for the number of hidden layers.
+        width_key: Full trial parameter name for the first hidden-layer width.
+        max_layers: Maximum number of hidden layers (default 4).
+
+    Returns:
+        The list of hidden-layer sizes (the funnel architecture).
+    """
+    num_layers = trial.suggest_int(layers_key, 1, max_layers, log=False)
+
+    # First hidden layer scaled to the data with a generous floor so small
+    # datasets still get a usable trunk.
+    min_hidden = max(64, input_dim // 2)
+    max_hidden = max(min_hidden, input_dim * 2)
+    step = max(32, input_dim // 16)
+    # Ensure max_hidden is reachable from min_hidden with the chosen step.
+    max_hidden = min_hidden + ((max_hidden - min_hidden) // step) * step
+
+    first_hidden = trial.suggest_int(width_key, min_hidden, max_hidden, step=step, log=False)
+
+    # Funnel: each subsequent layer halves the previous width (floored at 32).
+    hidden_dims = [first_hidden]
+    for i in range(1, num_layers):
+        hidden_dims.append(max(32, first_hidden // (2**i)))
+    return hidden_dims
